refactor(representation): clean up code in video components
