--- src/app/workspace/resource/representation/video/matrix.service.spec.ts
@@ -1,16 +1,0 @@
-import { TestBed } from '@angular/core/testing';
-
-import { MatrixService } from './matrix.service';
-
-describe('MatrixService', () => {
-    let service: MatrixService;
-
-    beforeEach(() => {
-        TestBed.configureTestingModule({});
-        service = TestBed.inject(MatrixService);
-    });
-
-    it('should be created', () => {
-        expect(service).toBeTruthy();
-    });
-});
